# [Vectorized][HashJoin] Opt HashJoin Performance