--- a/be/src/vec/columns/column_complex.h
+++ b/be/src/vec/columns/column_complex.h
@@ -153,9 +153,15 @@ public:
     void insert_indices_from(const IColumn& src, const int* indices_begin,
                              const int* indices_end) override {
         const Self& src_vec = assert_cast<const Self&>(src);
-        data.reserve(size() + (indices_end - indices_begin));
-        for (auto x = indices_begin; x != indices_end; ++x) {
-            data.push_back(src_vec.get_element(*x));
+        auto new_size = indices_end - indices_begin;
+
+        for (int i = 0; i < new_size; ++i) {
+            auto offset = *(indices_begin + i);
+            if (offset == -1) {
+                data.emplace_back(T{});
+            } else {
+                data.emplace_back(src_vec.get_element(offset));
+            }
         }
     }