# Proposed changes

Issue Number: close #xxx

## Problem Summary:

1. Opt hash join performance
Reduce mapped(RowRef) memory usage，the hash map traverse mapped is time consuming，so delete the block pointer（8 Bytes）is cache friendly.
After opt, the hash join performance is greatly improved。

My test:(SSB benchmark)
set runime_filter_type=0;
set parallel_fragment_exec_instance_num = 1;

inner join:
(before opt)
MySQL [ssb1]> SELECT count(c_custkey) FROM lineorder,customer WHERE lo_custkey = c_custkey;
+--------------------+
| count(c_custkey) |
+--------------------+
| 600037902 |
+--------------------+
1 row in set (43.25 sec)
(after opt)
MySQL [ssb1]> SELECT count(c_custkey) FROM customer inner join lineorder on lo_custkey = c_custkey;
+--------------------+
| count(`c_custkey`) |
+--------------------+
|          600037902 |
+--------------------+
1 row in set (21.66 sec)

ssb q3.3
MySQL [ssb1]> SELECT c_city, s_city, d_year, SUM(lo_revenue) AS REVENUE FROM customer, lineorder, supplier, dates WHERE lo_custkey = c_custkey AND lo_suppkey = s_suppkey AND lo_orderdate = d_datekey AND (c_city='UNITED KI1' OR c_city='UNITED KI5') AND (s_city='UNITED KI1' OR s_city='UNITED KI5') AND d_year >= 1992 AND d_year <= 1997 GROUP BY c_city, s_city, d_year ORDER BY d_year ASC, REVENUE DESC;
(before opt)
24 rows in set (32.15 sec)
(after opt):
24 rows in set (11.95 sec)

## Checklist(Required)

1. Does it affect the original behavior: (No)
2. Has unit tests been added: (No Need)
3. Has document been added or modified: (No Need)
4. Does it need to update dependencies: (No)
5. Are there any changes that cannot be rolled back: (No)

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...